refactor(noon-enabled): move to noon sub-folder
